backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai==1.10.0
qdrant-client==1.12.0
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-magic-bin==0.4.14
pydantic==2.5.3
pydantic-settings==2.1.0
cloudinary==1.38.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3
apscheduler==3.10.4
